gitignore, blueprint, create app
