+++ .gitignore
@@ -1,0 +1,4 @@
+Lib/
+Scripts/
+pyvenv.cfg
+__pycache__
