+++ web/app.py
@@ -1,33 +1,28 @@
-import json
 
-from bson import json_util
-from flask import Flask, request
-import dummy_returner
-import db
+import views
+from werkzeug.exceptions import HTTPException
 
-app = Flask(__name__)
+from flask import Flask, jsonify
 
 
-@app.route("/describe", methods=["POST"])
-def upload():
-    video_id = str(db.add_new_video(request.json))
-    # //DO SOME STUFF WITH AI
-    response = {'summary_score': dummy_returner.summary_score, 'summary_text': dummy_returner.summary_text,
-                'timestamp_generated': str(dummy_returner.timestamp_generated), 'video_id': str(video_id)}
-    db.create_result_record(parse_json(response))
-    return response
+def register_error_handlers(app):
+    @app.errorhandler(code_or_exception=HTTPException)
+    def handle_error(error):
+        return jsonify(error.description), error.code
 
 
-def parse_json(data):
-    return json.loads(json_util.dumps(data))
+def register_blueprints(app):
+    app.register_blueprint(blueprint=views.bp)
 
 
-@app.route("/find", methods=["GET"])
-def find_results_by_title():
-    args = request.args
-    results = db.get_search_results(args.get("title"))
-    return parse_json(results)
+def create_app():
+    app = Flask(import_name=__name__)
+    register_error_handlers(app=app)
+    register_blueprints(app=app)
+    
+    return app
 
 
 if __name__ == '__main__':
-    app.run()
+    app = create_app()
+    app.run(debug=True)
